feat: 添加待办列表 fix: 修改错误重启逻辑

+++ frontend/src/main.ts
@@ -17,16 +17,23 @@ const pinia = createPinia();
 app.config.errorHandler = (err, instance, info) => {
     console.error(err, instance, info);
 
-    showNotify({
-        type: 'danger',
-        message: '检测到程序错误，尝试清除本地数据并刷新页面！',
-        color: '#ad0000',
-        background: '#ffe1e1',
-    });
-    setTimeout(() => {
-        localStorage.clear();
-        window.location.reload();
-    }, 1000);
+    if (!!localStorage.key(0)) {
+        showNotify({
+            type: 'danger',
+            message: '检测到程序错误，尝试清除本地数据并刷新页面！',
+            color: '#ad0000',
+            background: '#ffe1e1',
+        });
+        setTimeout(() => {
+            localStorage.clear();
+            window.location.reload();
+        }, 1000);
+    } else {
+        showNotify({
+            type: 'danger',
+            message: '遇到无法恢复的错误，请联系管理员！',
+        });
+    }
 };
 
 pinia.use(piniaPersist);
@@ -42,11 +49,11 @@ setNotifyDefaultOptions({
     duration: 2000,
 });
 
+const settingStore = useSettingStore();
+
 window
     .matchMedia('(prefers-color-scheme: dark)')
     .addEventListener('change', (evt: MediaQueryListEvent) => {
-        const settingStore = useSettingStore();
-
         settingStore.themeIndex === 2 &&
             settingStore.changeTheme(evt.matches ? 'dark' : 'light');
     });
